+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 
 ## If updating from version 1.0.x it will be nessasary to re save the configuration, see readme for syntax or use plugin config.
 
+## 1.1.4
+Update 
+- Added optional switch to override chargign schedules
+- Added device name to optional switches
+- Will show extimated charge until disconnected
+- Bummped dependancies
+- Code cleanup 
+
 ## 1.1.3
 Update 
 -	Added battery info for multiple cars. Will need to update config with new info
-- Added switch to reboot charger.
+- Added optional switch to reboot charger.
 
 ## 1.1.2
 Update